flask-backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
marshmallow==3.19.0
pytest==7.1.2
python-dotenv==0.19.2